[MINT-3654] Admin actions section (#2209)
* [Implement admin actions sections for both home page and hkc milestone library]

* [Fix pagination page number over use and out of sync bug]

* [Add scroll to]

* [Update test and snapshots]

* [Refactor scroll to at bottom and add new tests with snapshots]

* [Add test to old component]

* [clean up snapshot]

* [Update line height]
