+++ src/features/HelpAndKnowledge/home.tsx
@@ -1,10 +1,11 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
+import { useLocation } from 'react-router-dom';
 import { GridContainer, SummaryBox } from '@trussworks/react-uswds';
 
+import UswdsReactLink from 'components/LinkWrapper';
 import MainContent from 'components/MainContent';
 import PageHeading from 'components/PageHeading';
-import ScrollLink from 'components/ScrollLink';
 import { convertToLowercaseAndDashes } from 'utils/modelPlan';
 
 import KeyContactDirectory from './_components/KeyContactDirectory';
@@ -24,6 +25,71 @@ const JUMP_TO_LINKS_I18NKEYS = [
 
 export const HelpAndKnowledgeHome = () => {
   const { t } = useTranslation('helpAndKnowledge');
+
+  const { hash } = useLocation();
+
+  // This handles when there's # in url and scroll to section
+  useEffect(() => {
+    let timer: ReturnType<typeof setTimeout>;
+
+    if (hash) {
+      timer = setTimeout(() => {
+        const id = hash.replace('#', '');
+        const element = document.getElementById(id);
+        if (element) {
+          element.scrollIntoView({ behavior: 'smooth' });
+        }
+      }, 100);
+    }
+
+    return () => {
+      if (timer) clearTimeout(timer);
+    };
+  }, [hash]);
+
+  // This handles updating the url on scroll to match the section in view
+  useEffect(() => {
+    let rafId: number | null = null;
+
+    const targetIds = JUMP_TO_LINKS_I18NKEYS.map(key =>
+      convertToLowercaseAndDashes(t(key))
+    );
+
+    const updateUrlOnScroll = () => {
+      const reversedIds = [...targetIds].reverse();
+
+      const activeId = reversedIds.find(id => {
+        const element = document.getElementById(id);
+        if (!element) return false;
+        const rect = element.getBoundingClientRect();
+
+        const isVisibleAtTop = rect.top >= 0 && rect.top <= 80;
+        const isAtBottom =
+          window.innerHeight + window.scrollY >=
+          document.documentElement.scrollHeight - 100;
+
+        return isVisibleAtTop || (id === reversedIds[0] && isAtBottom);
+      });
+
+      if (activeId && window.location.hash !== `#${activeId}`) {
+        window.history.replaceState(null, '', `#${activeId}`);
+      }
+    };
+
+    const handleScroll = () => {
+      if (rafId !== null) return;
+      rafId = requestAnimationFrame(() => {
+        updateUrlOnScroll();
+        rafId = null;
+      });
+    };
+
+    window.addEventListener('scroll', handleScroll, { passive: true });
+    return () => {
+      if (rafId !== null) cancelAnimationFrame(rafId);
+      window.removeEventListener('scroll', handleScroll);
+    };
+  }, [t]);
 
   return (
     <MainContent>
@@ -46,7 +112,13 @@ export const HelpAndKnowledgeHome = () => {
             </p>
             {JUMP_TO_LINKS_I18NKEYS.map((linkI18nkey, index) => (
               <div className="display-inline" key={linkI18nkey}>
-                <ScrollLink scrollTo={t(linkI18nkey)} hasIcon={false} />
+                <UswdsReactLink
+                  key={linkI18nkey}
+                  to={`#${convertToLowercaseAndDashes(t(linkI18nkey))}`} // 👈 Just the hash for same-page jumping
+                  className="usa-link"
+                >
+                  {t(linkI18nkey)}
+                </UswdsReactLink>
                 {index < JUMP_TO_LINKS_I18NKEYS.length - 1 && (
                   <div className="display-inline height-full width-1px border-left border-width-1px border-base-light margin-x-2" />
                 )}
